Add support for Azure CLI 2.0 commands
also remove some functions not dependent on root scope from root controller.
