+++ ng/scriptGenerators/ARMUrlParser.ts
@@ -1,0 +1,119 @@
+﻿module armExplorer {
+
+    enum ARMUrlParts {
+        Protocol,
+        Blank,
+        DomainName,
+        SubscriptionsKey,
+        SubscriptionsValue,
+        ResourceGroupsKey,
+        ResourceGroupsValue, // Urls with length <= ResourceGroupsValue will only include resourceId
+        ProviderKey,
+        ProviderValue, // start of resourcetype followed by 9(ResourceType1Name) , 11, 13, ... ( 8, 9 will always exist, 11, 13.. is optional)
+        ResourceType1Name,
+        ResourceType1Value // start of resourcename followed by 12, 14,...(12, 14,... is optional)
+    }
+
+    export class ARMUrlParser {
+        private originalUrl: string = "";
+        private url: string = "";
+        private urlParts: Array<string> = [];
+
+        constructor(private value: ISelectHandlerReturn, private actions: IAction[]) {
+            this.url = value.url;
+            this.originalUrl = value.url;
+            if (this.isSecureGet(this.value.httpMethod, this.url)) {
+                this.url = this.value.url.replace("/list", ""); // ignore /list since it is not part of resource url
+            }
+            this.urlParts = this.url.split("/");
+        }
+
+        private isSecureGet(httpMethod: string, url: string): boolean {
+            // sample url "https://management.azure.com/subscriptions/0000000-0000-0000-0000-0000000000000/resourceGroups/rgrp1/providers/Microsoft.Web/sites/WebApp120170517014339/config/appsettings/list"
+            return (httpMethod.toLowerCase() === HttpVerb[HttpVerb.POST].toLowerCase()) && url.toLowerCase().endsWith("/list")
+        }
+
+        private hasResourceType(): boolean {
+            return this.urlParts.length > ARMUrlParts.ResourceType1Name;
+        }
+
+        private hasResourceName(): boolean {
+            return this.urlParts.length > ARMUrlParts.ResourceType1Value;
+        }
+
+        getSubscriptionId(): string {
+            return this.urlParts[ARMUrlParts.SubscriptionsValue];
+        }
+
+        getResourceGroup(): string {
+            return this.urlParts[ARMUrlParts.ResourceGroupsValue];
+        }
+
+        getAPIVersion(): string {
+            return this.value.resourceDefinition.apiVersion;
+        }
+
+        getURL(): string {
+            return this.value.url;
+        }
+
+        getResourceDefinitionChildren(): string | string[] {
+            return this.value.resourceDefinition.children;
+        }
+
+        getOriginalURL(): string {
+            return this.originalUrl;
+        }
+
+        getHttpMethod(): string {
+            return this.value.httpMethod;
+        }
+
+        getActions(): IAction[] {
+            return this.actions;
+        }
+
+        getResourceActions(): string[] {
+            return this.value.resourceDefinition.actions;
+        }
+
+        hasResourceProvider(): boolean {
+            return this.urlParts.length > ARMUrlParts.ProviderKey;
+        }
+
+        isResourceGroupURL(): boolean {
+            return this.urlParts.length === (ARMUrlParts.ResourceGroupsKey + 1);
+        }
+
+        getResourceIdentifier(): ResourceIdentifier {
+            let resourceIdentifier = {} as ResourceIdentifier;
+
+            if (!this.hasResourceType()) {
+                // We only have resource Id
+                resourceIdentifier.resourceIdentifierType = ResourceIdentifierType.WithIDOnly;
+                resourceIdentifier.resourceId = this.url.replace("https://management.azure.com", "");
+            }
+            else {
+                resourceIdentifier.resourceGroup = this.urlParts[ARMUrlParts.ResourceGroupsValue];
+                resourceIdentifier.resourceType = this.urlParts[ARMUrlParts.ProviderValue] + "/" + this.urlParts[ARMUrlParts.ResourceType1Name];
+                resourceIdentifier.resourceIdentifierType = ResourceIdentifierType.WithGroupType;
+
+                if (this.hasResourceName()) {
+                    resourceIdentifier.resourceName = this.urlParts[ARMUrlParts.ResourceType1Value];
+                    resourceIdentifier.resourceIdentifierType = ResourceIdentifierType.WithGroupTypeName;
+                }
+                // check for resource sub types
+                for (let i = ARMUrlParts.ResourceType1Value + 1; i < this.urlParts.length; i++) {
+                    if (i % 2 == 1) {
+                        resourceIdentifier.resourceType += ("/" + this.urlParts[i]);
+                    }
+                    else {
+                        resourceIdentifier.resourceName += ("/" + this.urlParts[i]);
+                    }
+                }
+            }
+            return resourceIdentifier;
+        }
+    }
+
+}
