+++ ng/scriptGenerators/cli/resources/SubscriptionLocations.ts
@@ -1,0 +1,28 @@
+﻿module armExplorer {
+
+    export class SubscriptionLocations extends GenericResource {
+
+        constructor(resolver: ScriptParametersResolver) {
+            super(resolver);
+        }
+
+        getScript(action: ResourceAction): String {
+            let script: string = "";
+            switch (action) {
+            case ResourceAction.Get:
+                script = "az account list-locations\n\n";
+                break;
+            case ResourceAction.Invoke:
+            case ResourceAction.InvokeAction:
+            case ResourceAction.Set:
+            case ResourceAction.New:
+            case ResourceAction.NewResourceGroup:
+            case ResourceAction.RemoveAction:
+            default:
+                break;
+            }
+            return script === "" ? super.getScript(action) : script;
+        }
+    }
+
+}
