+++ ng/tsconfig.json
@@ -7,29 +7,47 @@
         "outFile": "manageWithTests.js",
         "listFiles": false
     },
-        "files": [
-        "PowerShellScriptGenerator.ts",
-        "models/IAction.ts",
-        "models/IArmTreeScope.ts",
-        "models/IMetadataObject.ts",
-        "models/IResourceDefinition.ts",
-        "models/IResourceSearch.ts",
-        "models/ISelectHandlerReturn.ts",
-        "models/ITenantDetails.ts",
-        "models/ITreeBranch.ts",
-        "models/ITreeControl.ts",
-        "polyfill.ts",
-        "typings/ace/ace.d.ts",
-        "typings/angularjs/angular-animate.d.ts",
-        "typings/angularjs/angular-cookies.d.ts",
-        "typings/angularjs/angular-mocks.d.ts",
-        "typings/angularjs/angular-resource.d.ts",
-        "typings/angularjs/angular-route.d.ts",
-        "typings/angularjs/angular-sanitize.d.ts",
-        "typings/angularjs/angular.d.ts",
-        "typings/cryptojs/cryptojs.d.ts",
-        "typings/jquery.cookie/jquery.cookie.d.ts",
-        "typings/jquery/jquery.d.ts",
-        "test/PowerShellScriptTests.ts"
-    ]
+  "files": [
+    "models/IAction.ts",
+    "models/IArmTreeScope.ts",
+    "models/IMetadataObject.ts",
+    "models/IResourceDefinition.ts",
+    "models/IResourceSearch.ts",
+    "models/ISelectHandlerReturn.ts",
+    "models/ITenantDetails.ts",
+    "models/ITreeBranch.ts",
+    "models/ITreeControl.ts",
+    "typings/ace/ace.d.ts",
+    "typings/angularjs/angular-animate.d.ts",
+    "typings/angularjs/angular-cookies.d.ts",
+    "typings/angularjs/angular-mocks.d.ts",
+    "typings/angularjs/angular-resource.d.ts",
+    "typings/angularjs/angular-route.d.ts",
+    "typings/angularjs/angular-sanitize.d.ts",
+    "typings/angularjs/angular.d.ts",
+    "typings/cryptojs/cryptojs.d.ts",
+    "typings/jquery.cookie/jquery.cookie.d.ts",
+    "typings/jquery/jquery.d.ts",
+    "common/polyfill.ts",
+    "common/helpers.ts",
+    "scriptGenerators/cli/resources/GernericResource.ts",
+    "scriptGenerators/cli/resources/ResourceGroup.ts",
+    "scriptGenerators/cli/resources/ResourceGroups.ts",
+    "scriptGenerators/cli/resources/Subscription.ts",
+    "scriptGenerators/cli/resources/SubscriptionLocations.ts",
+    "scriptGenerators/cli/resources/Subscriptions.ts",
+    "scriptGenerators/cli/resources/WebApp.ts",
+    "scriptGenerators/cli/resources/WebApps.ts",
+    "scriptGenerators/cli/CliScriptGenerator.ts",
+    "scriptGenerators/cli/ResourceHandlerResolver.ts",
+    "scriptGenerators/powershell/PowerShellScriptGenerator.ts",
+    "scriptGenerators/ARMUrlParser.ts",
+    "scriptGenerators/scriptGeneratorCommon.ts",
+    "scriptGenerators/ScriptParametersResolver.ts",
+    "test/TestCommon.ts",
+    "test/ARMUrlParserTests.ts",
+    "test/AzureCLIScriptTests.ts",
+    "test/PowerShellScriptTests.ts",
+    "test/ScriptParametersResolverTests.ts"
+  ]
 }
