+++ ng/views/AzureCLITab.html
@@ -1,0 +1,2 @@
+﻿<pre class="ng-binding">Please install the latest Azure PowerShell from <a href="https://docs.microsoft.com/en-us/cli/azure/install-azure-cli">here</a>.</pre>
+<div id="azurecli-editor" class="explorer-editor">></div>
