+++ ng/views/documentationTab.html
@@ -1,0 +1,12 @@
+﻿<table class="table table-hover table-bordered">
+    <thead>
+    <th>Property</th>
+    <th>Documentation</th>
+    </thead>
+    <tbody>
+        <tr ng-repeat="e in selectedResource.doc">
+            <th>{{ e.name }}</th>
+            <td>{{ e.doc }}</td>
+        </tr>
+    </tbody>
+</table>
