+++ ng/views/invokeActionResult.html
@@ -1,0 +1,14 @@
+﻿<div class="row">
+    <div class="col-lg-12">
+        <h1 class="page-header">{{ treeControl.get_selected_branch().label }}</h1>
+    </div>
+    <span class="loading-spinner" ng-show="loading || invoking">
+        <i class="fa fa-spinner fa-spin" title="Please wait" style="margin-top: 10px"></i>
+    </span>
+    <span id="success-marker" class="loading-spinner" style="display: none">
+        <i class="fa fa-check" title="success" style="margin-top: 10px"></i>
+    </span>
+    <span id="failure-marker" class="loading-spinner" style="display: none">
+        <i class="fa fa-exclamation-triangle" title="error" style="margin-top: 10px"></i>
+    </span>
+</div>
