+++ ng/views/powershellTab.html
@@ -1,0 +1,2 @@
+﻿<pre class="ng-binding">Please install the latest Azure PowerShell from <a href="https://github.com/Azure/azure-powershell/releases">here</a>.</pre>
+<div id="powershell-editor" class="explorer-editor">></div>
